+++ cloud/tests/unittests/multimodal/test_multimodal.py
@@ -4,77 +4,6 @@ import zipfile
 from autogluon.cloud import MultiModalCloudPredictor
 
 
-def test_multimodal_text_only(test_helper):
-    train_data = 'text_train.csv'
-    tune_data = 'text_tune.csv'
-    test_data = 'text_test.csv'
-    with tempfile.TemporaryDirectory() as root:
-        test_helper.prepare_data(train_data, tune_data, test_data)
-        time_limit = 60
-
-        predictor_init_args = dict(
-            label='label',
-            eval_metric='acc'
-        )
-        predictor_fit_args = dict(
-            train_data=train_data,
-            tuning_data=tune_data,
-            time_limit=time_limit
-        )
-        cloud_predictor = MultiModalCloudPredictor(
-            cloud_output_path='s3://autogluon-cloud-ci/test-multimodal-text',
-            local_output_path='test_multimodal_text_cloud_predictor'
-        )
-        cloud_predictor_no_train = MultiModalCloudPredictor(
-            cloud_output_path='s3://autogluon-cloud-ci/test-multimodal-text-no-train',
-            local_output_path='test_multimodal_text_cloud_predictor_no_train'
-        )
-        test_helper.test_functionality(
-            cloud_predictor,
-            predictor_init_args,
-            predictor_fit_args,
-            cloud_predictor_no_train,
-            test_data,
-            fit_instance_type='ml.g4dn.2xlarge'
-        )
-
-
-def test_multimodal_image_only(test_helper):
-    train_data = 'image_train_relative.csv'
-    train_image = 'shopee-iet.zip'
-    test_data = 'test_images/BabyPants_1035.jpg'
-    with tempfile.TemporaryDirectory() as root:
-        test_helper.prepare_data(train_data, train_image, test_data)
-        test_data = 'BabyPants_1035.jpg'
-        time_limit = 60
-
-        predictor_init_args = dict(
-            label='label',
-            eval_metric='acc'
-        )
-        predictor_fit_args = dict(
-            train_data=train_data,
-            time_limit=time_limit
-        )
-        cloud_predictor = MultiModalCloudPredictor(
-            cloud_output_path='s3://autogluon-cloud-ci/test-multimodal-image',
-            local_output_path='test_multimodal_image_cloud_predictor'
-        )
-        cloud_predictor_no_train = MultiModalCloudPredictor(
-            cloud_output_path='s3://autogluon-cloud-ci/test-multimodal-image-no-train',
-            local_output_path='test_multimodal_image_cloud_predictor_no_train'
-        )
-        test_helper.test_functionality(
-            cloud_predictor,
-            predictor_init_args,
-            predictor_fit_args,
-            cloud_predictor_no_train,
-            test_data,
-            image_path='shopee-iet.zip',
-            fit_instance_type='ml.g4dn.2xlarge',
-        )
-
-
 def test_multimodal_tabular_text(test_helper):
     train_data = 'tabular_text_train.csv'
     test_data = 'tabular_text_test.csv'
